fix(loop): fix goal mode gate architecture and scope filtering
Root cause: _filter_by_scope returned inactive mode-specific handlers
(e.g. mission) when no mode scope was active. The mission handler's
default STOP short-circuited the universal handler, preventing
GoalMode gates from ever being evaluated.

Additionally, GoalTurnGate now returns STOP when session.active is
False (set by update_goal tool), providing proper goal completion
detection.

Changes:
- Fix _filter_by_scope to exclude inactive mode-specific handlers
- GoalTurnGate checks session.active==False → returns STOP
- Rename GoalIterationGate → GoalTurnGate, inherit LoopGate directly
- Split goal_mode.py (825→~350 lines) into gates/prompts/helpers
- StandaloneRubricGate now inherits StopGate properly
- Unified handler registration via handler_registry.py
- Add logger.debug to 6 bare except blocks
- Initialize _gate_pending_stop/_gate_pending_continue in __init__
- Document _compute_similarity precondition (len(window) >= 2)
